+++ packages/quantic/force-app/main/default/lwc/quanticResultList/quanticResultList.js
@@ -23,7 +23,7 @@ export default class QuanticResultList extends LightningElement {
   }
 
   /**
-   * @param {import("coveo").Engine} engine
+   * @param {import("coveo").SearchEngine} engine
    */
   @api
   initialize(engine) {